requirements.txt
asgiref==3.4.1
Django==3.2.5cd
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
